Streamed audio dynamically instead of forcing it
This does prevent the delay in audio at the start of the exe version and prevents the audio passing in the web version from blocking everything arbitrarily as soon as you press. There is a click when the maze is regenerated in the web version and processing is done there but this is not noticeable.